refactor, sanitize calculator input
