+++ app/src/main/java/com/ahugenb/hra/navigation/NavScreen.kt
@@ -1,0 +1,7 @@
+package com.ahugenb.hra.navigation
+
+enum class NavScreen(val title: String) {
+    SCREEN_LIST("screenList"),
+    SCREEN_CALCULATOR("screenCalculator"),
+    SCREEN_PHILOSOPHY("screenPhilosophy")
+}
